CommandData Data inner class pulled out into separate class
